BAMParser Extensibility Enhancements
Created a generic version of the BAMParser class that will allow for
easier extension of the class by developers using the .NET Bio library.
Also, to maintain compatibility with existing code, there is also a
non-generic version of the BAMParser class that inherits from the
generic version of the class with the template parameter set to
SAMAlignedSequence.