Cleanup the MUMmerAligner Class
The MUMmer aligner exposed a PairWiseAlgorithm to be used to stitching
together of different seeds. There were however a number of constraints
that this algorithm must obey but which were not controlled for here.
In particular, the algorithm must be a global aligner that returns only
one alignment per sequence pair, but this was not enforced.  As we only
have the NeedleManWunsch aligner that fits the global bill here, I
enforced that this class must use this aligner for stitching together
seeds.  This is a breaking API change, but if anyone was using the old
way their code would silently be giving bad results, so it is an
appropriate change.

This change also added comments for the aligner.

Finally, it also fixes an issue where the insertions were not
tallied correctly as it was converting to List<int> instead of
List<long>