Less Memory Usage / More Comments in NeedlemanWunschAligner
For these dynamic programming problems, when traversing through the
matrix one must store the entire traceback matrix,
but the scores are only needed for the current row and the row above it.

In the current implementation, we only use two rows for the "top
scoring" (and match) matrix, but fill out entire matrices for the
deletion and insert (or vertical and horizontal) matrices.  This is
inefficient, as we then need 2 x M x N memory rather than 2 x 2 x N
memory.  I switched this to only use the smaller amount.  Additionally:

* Changed the minimum score on the edges from an arbitray low number to a constant near Int32.MinValue for clarity.

* Added more comments throughout.

* Rather than have three arrays, I used 1 array of a struct which may help memory locality for very large global alignments.

* Added more unit tests that directly compare against EMBOSS Needle alignment program, found and reported a bug in NEEDLE while doing that...

* Change all the vague category "Priority2" labels in NW tests to a newer version.