Allow Redundant Paths to Consider Indels
Before the redundant path purger could only remove SNPs, this allows
it to call paths like A -> B -> F / A->C->D->E->F as redundant,
simplifying the graph
quite a bit.

Changes include:

* Make PathWithOrientation class to be internal to the assembly
* Remove redundant null check on pass graph
* Allow paths to converge if they have different lengths but wind up at
same end node
* Verified that the paths not only end on the same node, but also
approach it from the same direction

The checking for loops needs to be much faster than it is.